Fix reading from rdisk devices.
Reading the logical length (if it's less than the last extent) rather than the whole extent can reduce unnecessary I/O from slow devices, but reading *exactly* the logical length won't work if it isn't a multiple of kISOStandardBlockSize and we're reading from an rdisk device where that's mandatory. So, round up to the volume's block size, which also must be a multiple of kISOStandardBlockSize.