Starting work on #7, Apple Partition Map support. The first step is to decouple the subcommands from assuming their input is HFS or is the whole input file.
This adds a new class, ImpVolumeProbe, that reads an input file and attempts to discover volumes in it. It can discover HFS or HFS+ volumes and will call a block with each one found.

convert and extract ignore any volumes that aren't HFS; convert also only converts the first HFS volume. list and analyze already supported HFS and HFS+, and that logic is now simpler.