Put the catalog builder in charge of determining how many nodes are needed in the catalog file. This completely prevents another case like the one we encountered with Doom II (see #18) where conversion could fail due to the catalog file being bigger than estimated—now, the allocation for the catalog file will always be just enough to hold it.
The extents overflow file's node count is now hard-coded at 2 in the defragmenting converter. Less than ideal, but should be fine since I don't think the defragmenting converter will ever touch the extents overflow file. A file would either need to be fragmented (impossible with that converter) or bigger than we're likely to ever see on an HFS volume, especially with the logical and physical length fields being 32-bit there.